+++ semgrep-core/src/parsing/pfff/java_to_generic.ml
@@ -421,6 +421,19 @@ and arguments v : G.argument list G.bracket = bracket (list argument) v
 
 and fix_op v = H.conv_incr v
 
+and resource (v : resource) : G.stmt =
+  match v with
+  | Left v ->
+      let ent, v = var_with_init v in
+      G.DefStmt (ent, G.VarDef v) |> G.s
+  | Right e -> G.ExprStmt (expr e, unsafe_fake "") |> G.s
+
+and resources (t1, v, t2) =
+  match v with
+  | [] -> G.Block (t1, [], t2) |> G.s
+  | [ v ] -> resource v
+  | _ :: _ -> G.Block (t1, list resource v, t2) |> G.s
+
 and stmt st =
   match st with
   | EmptyStmt t -> G.Block (t, [], t) |> G.s
@@ -469,9 +482,12 @@ and stmt st =
   | Sync (v1, v2) ->
       let v1 = expr v1 and v2 = stmt v2 in
       G.OtherStmtWithStmt (G.OSWS_Sync, [ G.E v1 ], v2) |> G.s
-  | Try (t, _v0TODO, v1, v2, v3) ->
+  | Try (t, v0, v1, v2, v3) -> (
       let v1 = stmt v1 and v2 = catches v2 and v3 = option tok_and_stmt v3 in
-      G.Try (t, v1, v2, v3) |> G.s
+      let try_stmt = G.Try (t, v1, v2, v3) |> G.s in
+      match v0 with
+      | None -> try_stmt
+      | Some r -> G.WithUsingResource (t, resources r, try_stmt) |> G.s)
   | Throw (t, v1) ->
       let v1 = expr v1 in
       G.Throw (t, v1, G.sc) |> G.s